requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
appdirs==1.4.4
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioread @ file:///Users/runner/miniforge3/conda-bld/audioread_1759429664790/work
av==15.1.0
babel==2.17.0
backports.functools-lru-cache==2.0.0
backports.zstd @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_backports.zstd_1767045020/work
bangla==0.0.5
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761202941799/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.1.8
cloudpathlib==0.23.0
coloredlogs==15.0.1
colorlog==6.10.1
confection==0.1.5
contourpy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_contourpy_1769156091/work
coqpit==0.0.17
csvw==3.7.0
ctranslate2==4.6.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cycler_1764466758/work
cymem==2.0.13
Cython @ file:///Users/runner/miniforge3/conda-bld/cython_1767576781053/work
dateparser==1.1.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
deep-translator==1.11.4
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
Deprecated==1.3.1
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
fastapi==0.128.0
fastdtw==0.3.4
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1765632680641/work
frozenlist==1.8.0
fsspec==2025.12.0
g2pkk==0.1.2
gradio==4.15.0
gradio_client==0.8.1
groovy==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hangul-romanize==0.1.0
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
HyperPyYAML==1.2.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
indic_transliteration==2.3.76
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
jsonlines==1.2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
julius==0.2.7
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_kiwisolver_1762488860/work
language-tags==1.2.0
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
Levenshtein==0.27.3
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1741797566785/work
lightning==2.4.0
lightning-utilities==0.15.2
llvmlite==0.46.0
loguru==0.7.3
lru-dict==1.4.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1762504096247/work
multidict==6.7.0
munkres==1.1.4
murmurhash==1.0.15
mwparserfromhell==0.7.2
narwhals==2.15.0
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba @ file:///Users/runner/miniforge3/conda-bld/numba_1765808305328/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225640867/work/dist/numpy-1.26.4-cp311-cp311-macosx_11_0_arm64.whl#sha256=b62b8b824024855c3eb209bd45bfcf60bc9a11b94cbf9d5b7b25dc503021e672
omegaconf==2.3.0
onnxruntime==1.19.2
optuna==4.6.0
orjson==3.11.5
packaging==23.2
pandas==1.5.3
phonemizer==3.3.0
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1754941678315/work
preshed==3.0.12
primePy==1.3
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
py-espeak-ng==0.1.8
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
pynndescent==0.6.0
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-crfsuite==0.9.12
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.22
python-telegram-bot==22.6
pytorch-lightning==2.4.0
pytorch-metric-learning==2.9.0
pytz==2025.2
pyworld @ file:///Users/runner/miniforge3/conda-bld/pyworld_1761314301395/work
PyYAML==6.0.3
RapidFuzz==3.14.3
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
rfc3986==1.5.0
rich==14.2.0
roman==5.2
rpds-py==0.30.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.14.14
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_scikit-learn_1766550842/work/dist/scikit_learn-1.8.0-cp311-cp311-macosx_11_0_arm64.whl#sha256=e9bfee154265c2c47be943a5c346c467a67d1d1bed3bb034914a57d76df0c1a9
scipy==1.16.3
segments==2.3.0
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.5.0
sortedcontainers==2.4.0
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1737836266465/work
soupsieve==2.8.3
soxr @ file:///Users/runner/miniforge3/conda-bld/soxr-python_1762684782108/work
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.45
srsly==2.5.2
standard-aifc @ file:///Users/runner/miniforge3/conda-bld/standard-aifc_1761318476720/work
standard-sunau @ file:///Users/runner/miniforge3/conda-bld/standard-sunau_1761318525244/work
starlette==0.50.0
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.3.0
thinc==8.3.10
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
torch==2.5.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.8.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.38.2
TTS==0.22.0
typeguard==4.4.4
typer==0.20.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.11
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1763054903095/work
Unidecode==1.4.0
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
websockets==11.0.3
Werkzeug==3.1.5
whisperx==3.6.0
wikitextprocessor==0.4.96
wiktextract==1.99.7
wiktionaryparser==0.0.97
wrapt==2.0.1
yarl==1.22.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
